+++ CHANGES.md
@@ -5,6 +5,7 @@
   (warning 67)
 * Add support for the `s` element
 * Add support for the `bdi` element
+* Add support for the `search` element
 * Fix typo `whitout` in type definition
 	(#324 by Martin @MBodin Bodin)
 * Add support for the clip-path presentation attribute